Further test fixes
Revert the auth mock changes and instead merge the 2 problematic
test classes into one because they are identical in flow.

Fixes test issue.
