+++ src/test/kotlin/uk/gov/justice/digital/hmpps/prison/visitbooker/registry/integration/IntegrationTestBase.kt
@@ -1,6 +1,5 @@
 package uk.gov.justice.digital.hmpps.prison.visitbooker.registry.integration
 
-import com.github.tomakehurst.wiremock.WireMockServer
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.AfterAll
 import org.junit.jupiter.api.AfterEach
@@ -65,9 +64,9 @@ abstract class IntegrationTestBase {
     @BeforeAll
     @JvmStatic
     fun startMocks() {
-      prisonOffenderSearchMockServer.startIfNeeded()
-      visitSchedulerMockServer.startIfNeeded()
-      prisonerContactRegistryMockServer.startIfNeeded()
+      prisonOffenderSearchMockServer.start()
+      visitSchedulerMockServer.start()
+      prisonerContactRegistryMockServer.start()
     }
 
     @AfterAll
@@ -76,12 +75,6 @@ abstract class IntegrationTestBase {
       prisonOffenderSearchMockServer.stop()
       visitSchedulerMockServer.stop()
       prisonerContactRegistryMockServer.stop()
-    }
-
-    private fun WireMockServer.startIfNeeded() {
-      if (!isRunning) {
-        start()
-      }
     }
   }
 
@@ -122,7 +115,6 @@ abstract class IntegrationTestBase {
 
   @BeforeEach
   fun resetStubs() {
-    startMocks()
     entityHelper.deleteAll()
     prisonOffenderSearchMockServer.resetAll()
     visitSchedulerMockServer.resetAll()
