+++ src/test/kotlin/uk/gov/justice/digital/hmpps/prison/visitbooker/registry/integration/mock/HmppsAuthMockServer.kt
@@ -5,6 +5,7 @@ import com.github.tomakehurst.wiremock.client.WireMock.aResponse
 import com.github.tomakehurst.wiremock.client.WireMock.get
 import com.github.tomakehurst.wiremock.client.WireMock.post
 import com.github.tomakehurst.wiremock.client.WireMock.urlEqualTo
+import org.junit.jupiter.api.extension.AfterAllCallback
 import org.junit.jupiter.api.extension.BeforeAllCallback
 import org.junit.jupiter.api.extension.BeforeEachCallback
 import org.junit.jupiter.api.extension.ExtensionContext
@@ -14,6 +15,7 @@ import uk.gov.justice.digital.hmpps.prison.visitbooker.registry.integration.mock
 
 class HmppsAuthExtension :
   BeforeAllCallback,
+  AfterAllCallback,
   BeforeEachCallback {
 
   companion object {
@@ -22,11 +24,10 @@ class HmppsAuthExtension :
   }
 
   override fun beforeAll(context: ExtensionContext) {
-    hmppsAuthApi.startIfNeeded()
+    hmppsAuthApi.start()
   }
 
   override fun beforeEach(context: ExtensionContext) {
-    hmppsAuthApi.startIfNeeded()
     hmppsAuthApi.resetAll()
     hmppsAuthApi.stubGrantToken()
     hmppsAuthApi.stubGetUserDetails("created-user")
@@ -34,10 +35,8 @@ class HmppsAuthExtension :
     hmppsAuthApi.stubGetUserDetails("cancelled-user")
   }
 
-  private fun HmppsAuthMockServer.startIfNeeded() {
-    if (!isRunning) {
-      start()
-    }
+  override fun afterAll(context: ExtensionContext) {
+    hmppsAuthApi.stop()
   }
 }
